reformatted to replace relation with the new datasource
