+++ src/main/scala/com/databricks/spark/avro/AvroOutputWriterFactory.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.databricks.spark.avro.avroRelation2
+package com.databricks.spark.avro
 
 import org.apache.hadoop.mapreduce.TaskAttemptContext
 import org.apache.spark.sql.sources.{OutputWriter, OutputWriterFactory}
@@ -22,11 +22,11 @@ import org.apache.spark.sql.types.StructType
 
 
 private[avro] class AvroOutputWriterFactory(schema: StructType,
-                                            recordName: String,
-                                            recordNamespace: String) extends OutputWriterFactory {
+    recordName: String,
+    recordNamespace: String) extends OutputWriterFactory {
 
   override def newInstance(path: String,
-                           dataSchema: StructType,
-                           context: TaskAttemptContext): OutputWriter =
+      dataSchema: StructType,
+      context: TaskAttemptContext): OutputWriter =
     new AvroOutputWriter(path, context, schema, recordName, recordNamespace)
 }
