--- src/main/scala/com/databricks/spark/avro/avroRelation2/DefaultSource.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright 2014 Databricks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.databricks.spark.avro.avroRelation2
-
-import org.apache.spark.sql.SQLContext
-import org.apache.spark.sql.sources.{HadoopFsRelation, HadoopFsRelationProvider}
-import org.apache.spark.sql.types.StructType
-
-private[avro] class DefaultSource extends HadoopFsRelationProvider {
-
-  def createRelation(sqlContext: SQLContext,
-                     paths: Array[String],
-                     dataSchema: Option[StructType],
-                     partitionColumns: Option[StructType],
-                     parameters: Map[String, String]): HadoopFsRelation = {
-    new AvroRelation2(paths, dataSchema, partitionColumns, parameters)(sqlContext)
-  }
-}
-
