+++ src/test/scala/com/databricks/spark/avro/AvroReadBenchmark.scala
@@ -26,29 +26,14 @@ object AvroReadBenchmark {
 
     val startTime = System.nanoTime
     TestSQLContext
-      .avroFile(AvroFileGenerator.outputDir)
+      .read
+      .avro(AvroFileGenerator.outputDir)
       .select("string")
       .count()
     val endTime = System.nanoTime
     val executionTime = TimeUnit.SECONDS.convert(endTime - startTime, TimeUnit.NANOSECONDS)
 
     println(s"\n\n\nFinished benchmark test - result was $executionTime seconds\n\n\n")
-
-    println("\n\n\nStaring benchmark test with DataFrameReader - " +
-      "creating DataFrame from benchmark avro files\n\n\n")
-
-    val startTime2 = System.nanoTime
-    TestSQLContext
-      .read
-      .avro(AvroFileGenerator.outputDir)
-      .select("string")
-      .count()
-    val endTime2 = System.nanoTime
-    val executionTime2 = TimeUnit.SECONDS.convert(endTime2 - startTime2, TimeUnit.NANOSECONDS)
-
-    println(s"\n\n\nFinished benchmark test with DataFrameReader -" +
-      s" result was $executionTime2 seconds\n\n\n")
-
 
     TestSQLContext.sparkContext.stop()  // Otherwise scary exception message appears
   }
